Fixed small bug in the design creation.
Even when no numbers are to be presented trials were created with
numbers in them, but not actually presented. The problem is that this
placed a restriction on the model selection.

It also seems that with the current list of eligible letters that no
more then 7 letters max can be displayed.