Optimized design creation and finished presentation of 1 thru 8 letters.
Finalized 8 letter presentation using 15 letter pool. I tried with 14
letters but only sequences of 16 trials were possible. So I added the
letter P back to the pile and it can work now. I also had to modify how
designs are created based on the letters used for that trial and the
ones NOT used for the prevous trials. This allowed easier design matrix
selection and creation. Optimization of the design showed that a 3 3 3
design duration was much more efficient then a 3 2 3 design duration.
Now I need to fix an error in the instruction presentation.